Add meson build file
The current build files config.mk and Makefile do not support diffrent
library path. Therefore, here is an attempt to write a meson.build file
based on pkg-config.

It also supports generation of wayland protocol code using
wayland-scanner, as distributions are not budling wlroots protocol
files, it is included in this commit (wlr-layer-shell-unstable-v1.xml).